requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
requests==2.32.3
httpx==0.27.0
beautifulsoup4==4.12.3
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
pydantic==2.7.3
APScheduler==3.10.4
jinja2==3.1.4
openai==1.44.0
reportlab==4.2.2
